Move content of IKeyboardHandler to the KeybordHandlerBase class and rename files to the snake style
