+++ rosbag2_keyboard_handler/CMakeLists.txt
@@ -14,8 +14,8 @@ if(WIN32)
 endif()
 
 set(SOURCES
-    src/IKeyboardHandler.cpp
-    src/KeyboardHandlerUnixImpl.cpp
+    src/keyboard_handler_base.cpp
+    src/keyboard_handler_unix_impl.cpp
 )
 
 add_library(${PROJECT_NAME} SHARED ${SOURCES})
@@ -62,7 +62,7 @@ if(BUILD_TESTING)
   ament_lint_auto_find_test_dependencies()
 
   set(keyboard_handler_test_sources
-      test/KeyboardHandlerTests.cpp)
+      test/keyboard_handler_tests.cpp)
 
   ament_add_gmock(test_keyboard_handler ${keyboard_handler_test_sources})
   target_include_directories(test_keyboard_handler PUBLIC include)
